hwk3/requirements.txt
functions-framework>=3.6.0
google-cloud-storage>=2.14.0
